Extract domain mappers to create entity & pojo mappers
