+++ app/src/main/java/com/jaimegc/covid19tracker/data/datasource/CovidTrackerDatasources.kt
@@ -20,6 +20,15 @@ import com.jaimegc.covid19tracker.data.room.entities.RegionStatsEntity
 import com.jaimegc.covid19tracker.data.room.entities.SubRegionEntity
 import com.jaimegc.covid19tracker.data.room.entities.SubRegionStatsEntity
 import com.jaimegc.covid19tracker.data.room.entities.WorldStatsEntity
+import com.jaimegc.covid19tracker.data.room.mapper.mapEntity
+import com.jaimegc.covid19tracker.data.room.mapper.mapEntityMenuValid
+import com.jaimegc.covid19tracker.data.room.mapper.mapEntityValid
+import com.jaimegc.covid19tracker.data.room.mapper.toEntity
+import com.jaimegc.covid19tracker.data.room.mapper.toPojoCountriesOrdered
+import com.jaimegc.covid19tracker.data.room.mapper.toPojoRegionDomain
+import com.jaimegc.covid19tracker.data.room.mapper.toPojoRegionsOrdered
+import com.jaimegc.covid19tracker.data.room.mapper.toPojoSubRegionDomain
+import com.jaimegc.covid19tracker.data.room.mapper.toPojoSubRegionsOrdered
 import com.jaimegc.covid19tracker.domain.model.CountryOneStats
 import com.jaimegc.covid19tracker.domain.model.CovidTracker
 import com.jaimegc.covid19tracker.domain.model.DomainError
@@ -34,17 +43,8 @@ import com.jaimegc.covid19tracker.domain.model.ListSubRegionAndStats
 import com.jaimegc.covid19tracker.domain.model.ListSubRegionStats
 import com.jaimegc.covid19tracker.domain.model.ListWorldStats
 import com.jaimegc.covid19tracker.domain.model.RegionOneStats
-import com.jaimegc.covid19tracker.domain.model.mapEntity
-import com.jaimegc.covid19tracker.domain.model.mapEntityMenuValid
-import com.jaimegc.covid19tracker.domain.model.mapEntityValid
 import com.jaimegc.covid19tracker.domain.model.toDomain
-import com.jaimegc.covid19tracker.domain.model.toEntity
-import com.jaimegc.covid19tracker.domain.model.toPojoRegionDomain
 import com.jaimegc.covid19tracker.domain.model.toStatsDomain
-import com.jaimegc.covid19tracker.domain.model.toPojoCountriesOrdered
-import com.jaimegc.covid19tracker.domain.model.toPojoRegionsOrdered
-import com.jaimegc.covid19tracker.domain.model.toPojoSubRegionDomain
-import com.jaimegc.covid19tracker.domain.model.toPojoSubRegionsOrdered
 import com.jaimegc.covid19tracker.ui.base.states.MenuItemViewType
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
